[AAE-6389] removed duplicate services and refactored existing
