+++ lib/core/mock/public-api.ts
@@ -37,4 +37,7 @@ export * from './form/start-form.component.mock';
 export * from './form/form.service.mock';
 export * from './form/widget-visibility.service.mock';
 export * from './jwt-helper.service.spec';
+export * from './identity-group.mock';
+export * from './identity-user.mock';
 export * from './identity-group.service.mock';
+export * from './identity-user.service.mock';
